# projects/ungouge-app/services/api-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Ungouge.ai API Gateway
  description: Entry point for all backend requests - handles auth, routing, rate limiting
  version: 2.0.0
  contact:
    name: Ungouge.ai Support
    email: support@ungouge.ai
servers:
  - url: https://api.ungouge.ai
    description: Production
  - url: https://api-staging.ungouge.ai
    description: Staging
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/quotes/{quote_id}:
    get:
      tags: [quotes]
      summary: Get quote analysis results
      security:
        - BearerAuth: []
      parameters:
        - name: quote_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quote details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '404':
          description: Quote not found
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Quote:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [uploaded, processing, analyzed, paid, failed]
        file_url:
          type: string
          format: uri
        uploaded_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
        extracted_data:
          type: object
          properties:
            contractor:
              type: object
              properties:
                name:
                  type: string
                license:
                  type: string
                contact:
                  type: string
            line_items:
              type: array
              items:
                type: object
                properties:
                  description:
                    type: string
                  quantity:
                    type: number
                  unit:
                    type: string
                  unit_price:
                    type: number
                  total:
                    type: number
            totals:
              type: object
              properties:
                subtotal:
                  type: number
                tax:
                  type: number
                total:
                  type: number
        cost_analysis:
          type: object
          properties:
            total_cost:
              type: number
            fair_price_range:
              type: object
              properties:
                min:
                  type: number
                max:
                  type: number
            confidence_score:
              type: number
              minimum: 0
              maximum: 1
            breakdown:
              type: array
              items:
                type: object
                properties:
                  category:
                    type: string
                  estimated_cost:
                    type: number
                  quoted_cost:
                    type: number
                  variance:
                    type: number
            recommendations:
              type: array
              items:
                type: string